added vector support via IPC

+++ psi/include/psi/shared/ipc/IPCCall.h
@@ -35,7 +35,7 @@ struct IPCCall {
         offset += serializable::write(buffer + offset, m_method_id);
         offset += serializable::write(buffer + offset, m_client_id);
         offset += serializable::write(buffer + offset, m_cb_index);
-        uint64_t args_sz = 0;
+        uint16_t args_sz = 0;
         fnPerTuple(
             [&](const auto &arg) -> int {
                 serializer::serializeTuple(buffer + offset + sizeof(decltype(m_args_sz)), arg, args_sz);
